Move rest com to `src/communications`.
