+++ src/communications/classRestCom.py
